+++ src/main/groovy/fr/nereide/reference/common/UiLabelReference.groovy
@@ -17,9 +17,9 @@
 
 package fr.nereide.reference.common
 
+import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiReferenceBase
-import com.intellij.psi.xml.XmlAttributeValue
 import com.intellij.util.xml.DomElement
 import fr.nereide.project.ProjectServiceInterface
 import fr.nereide.project.utils.MiscUtils
@@ -28,6 +28,10 @@ import org.jetbrains.annotations.Nullable
 class UiLabelReference extends PsiReferenceBase<PsiElement> {
     UiLabelReference(PsiElement element) {
         super(element)
+    }
+
+    UiLabelReference(PsiElement element, TextRange range) {
+        super(element, range)
     }
 
     @Nullable