+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Fixed
+- [#2] Fixes references in xml labels in case there is not only a label in the attribute value.
+
 ## [1.5.8]
 ### Changed
 - [#3] Returns a multi reference for services in case of multiple implementations for a single service name.